Refactor: Organize SEO utilities into separate files
